refactor: merge dispatcher and listener and private methods
